added DotNetflixMobileAPI with GraphQL and updated solution to .net 8
