+++ DotNetflixServer/DockerfileDotNetflixAPI
@@ -1,6 +1,6 @@
 #See https://aka.ms/containerfastmode to understand how Visual Studio uses this Dockerfile to build your images for faster debugging.
 
-FROM mcr.microsoft.com/dotnet/aspnet:7.0 AS base
+FROM mcr.microsoft.com/dotnet/aspnet:8.0 AS base
 WORKDIR /app
 EXPOSE 80
 EXPOSE 443
@@ -8,7 +8,7 @@ EXPOSE 443
 ADD ./Domain/jsons /Domain/jsons
 ADD ./https/aspnetapp.pfx /https/aspnetapp.pfx
 
-FROM mcr.microsoft.com/dotnet/sdk:7.0 AS build
+FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
 WORKDIR /src
 COPY ["DotNetflixAPI/DotNetflixAPI.csproj", "DotNetflixAPI/"]
 RUN dotnet restore "DotNetflixAPI/DotNetflixAPI.csproj"
