+++ DotNetflixServer/DotNetflix.Admin.Application/DotNetflix.Admin.Application.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
+        <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
     </PropertyGroup>
@@ -19,7 +19,7 @@
       <PackageReference Include="DotNetflixOAuth" Version="1.0.1" />
       <PackageReference Include="FluentValidation" Version="11.8.0" />
       <PackageReference Include="FluentValidation.DependencyInjectionExtensions" Version="11.8.0" />
-      <PackageReference Include="Microsoft.Extensions.Http" Version="7.0.0" />
+      <PackageReference Include="Microsoft.Extensions.Http" Version="8.0.0" />
     </ItemGroup>
 
 </Project>
