+++ DotNetflixServer/DotNetflix.Application/DotNetflix.Application.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
+        <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
     </PropertyGroup>
@@ -20,7 +20,7 @@
       <PackageReference Include="FluentValidation" Version="11.8.0" />
       <PackageReference Include="FluentValidation.DependencyInjectionExtensions" Version="11.8.0" />
       <PackageReference Include="MediatR" Version="12.1.1" />
-      <PackageReference Include="Microsoft.Extensions.Identity.Core" Version="7.0.5" />
+      <PackageReference Include="Microsoft.Extensions.Identity.Core" Version="8.0.0" />
     </ItemGroup>
 
 </Project>
