+++ DotNetflixServer/DotNetflixMobileAPI/DotNetflixMobileAPI.csproj
@@ -1,0 +1,24 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+    <PropertyGroup>
+        <TargetFramework>net8.0</TargetFramework>
+        <Nullable>enable</Nullable>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
+    </PropertyGroup>
+    
+    <ItemGroup>
+        <PackageReference Include="Grpc.AspNetCore" Version="2.61.0" />
+        <PackageReference Include="GraphQL.Server.Transports.AspNetCore" Version="7.7.1" />
+        <PackageReference Include="HotChocolate.AspNetCore" Version="13.9.0" />
+        <PackageReference Include="HotChocolate.Data.EntityFramework" Version="13.9.0" />
+        <PackageReference Include="MediatR" Version="12.1.1" />
+    </ItemGroup>
+    
+    <ItemGroup>
+        <ProjectReference Include="..\API.Shared\API.Shared.csproj" />
+        <ProjectReference Include="..\DotNetflix.Application\DotNetflix.Application.csproj" />
+        <ProjectReference Include="..\Services.Infrastructure\Services.Infrastructure.csproj" />
+    </ItemGroup>
+
+</Project>
