+++ DotNetflixServer/DotNetflixMobileAPI/appsettings.json
@@ -1,0 +1,22 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "DefaultConnection": ""
+  },
+  "SAPassword": "",
+  "SmtpSetting": {
+    "FromName": ".Netflix",
+    "FromAddress": "oristask@gmail.com",
+    "MailServerAddress": "smtp.gmail.com",
+    "FromPassword": "",
+    "MailServerPort": 25
+  },
+  "StorageApiBaseUrl": "",
+  "PaymentGrpcAddress": ""
+}
