+++ DotNetflixServer/DotNetflixServer.sln
@@ -33,6 +33,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetflix.Storage", "DotNe
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetflix.Payment", "DotNetflix.Payment\DotNetflix.Payment.csproj", "{732E9DE0-3B83-4EE3-898D-3B4F5744635B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetflixMobileAPI", "DotNetflixMobileAPI\DotNetflixMobileAPI.csproj", "{34C4802C-4794-4A0C-9529-F91E5A00E10B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "API.Shared", "API.Shared\API.Shared.csproj", "{B333C89F-B978-4ADD-A780-C4CD56567846}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -95,6 +99,14 @@ Global
 		{732E9DE0-3B83-4EE3-898D-3B4F5744635B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{732E9DE0-3B83-4EE3-898D-3B4F5744635B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{732E9DE0-3B83-4EE3-898D-3B4F5744635B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{34C4802C-4794-4A0C-9529-F91E5A00E10B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{34C4802C-4794-4A0C-9529-F91E5A00E10B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{34C4802C-4794-4A0C-9529-F91E5A00E10B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{34C4802C-4794-4A0C-9529-F91E5A00E10B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B333C89F-B978-4ADD-A780-C4CD56567846}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B333C89F-B978-4ADD-A780-C4CD56567846}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B333C89F-B978-4ADD-A780-C4CD56567846}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B333C89F-B978-4ADD-A780-C4CD56567846}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -107,5 +119,6 @@ Global
 		{47A40B6E-DD7E-4E6B-B196-4C0F9D8F94CC} = {FA5A2799-258C-4215-A32C-094E3FC4F44A}
 		{F18CF318-699D-4309-895A-BB1701C0D3BA} = {FA5A2799-258C-4215-A32C-094E3FC4F44A}
 		{55D630EC-1C17-429A-B944-D83B606697CB} = {FA5A2799-258C-4215-A32C-094E3FC4F44A}
+		{B333C89F-B978-4ADD-A780-C4CD56567846} = {FA5A2799-258C-4215-A32C-094E3FC4F44A}
 	EndGlobalSection
 EndGlobal
