+++ DotNetflixServer/Services.Infrastructure/Services.Infrastructure.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
+        <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
     </PropertyGroup>
@@ -10,7 +10,7 @@
       <PackageReference Include="DotNetflixOAuth" Version="1.0.1" />
       <PackageReference Include="Google.Apis.Auth" Version="1.60.0" />
       <PackageReference Include="IdentityPasswordGenerator" Version="1.0.5" />
-      <PackageReference Include="Microsoft.Extensions.Options" Version="7.0.1" />
+      <PackageReference Include="Microsoft.Extensions.Options" Version="8.0.0" />
     </ItemGroup>
 
     <ItemGroup>
